+++ BesLyric/BesLyric.vcxproj.filters
@@ -139,6 +139,12 @@
     <ClCompile Include="utility\UrlEncoding.cpp">
       <Filter>Source Files\utility</Filter>
     </ClCompile>
+    <ClCompile Include="entity\SearcherNetEaseCloud.cpp">
+      <Filter>Source Files\entity</Filter>
+    </ClCompile>
+    <ClCompile Include="utility\HttpRequest.cpp">
+      <Filter>Source Files\utility</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="BesLyric.rc">
